+++ spec/unit/rendezvous/simpleHttpTransport.spec.ts
@@ -18,7 +18,7 @@ import MockHttpBackend from "matrix-mock-request";
 
 import type { MatrixClient } from "../../../src";
 import { RendezvousFailureReason } from "../../../src/rendezvous";
-import { MSC3886SimpleHttpRendezvousTransport } from "../../../src/rendezvous/transports/simpleHttpTransport";
+import { MSC3886SimpleHttpRendezvousTransport } from "../../../src/rendezvous/transports";
 
 function makeMockClient(opts: { userId: string, deviceId: string, msc3886Enabled: boolean}): MatrixClient {
     return {
