+++ src/rendezvous/RendezvousError.ts
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { RendezvousFailureReason } from "./cancellationReason";
+import { RendezvousFailureReason } from ".";
 
 export class RendezvousError extends Error {
     constructor(message: string, public readonly code: RendezvousFailureReason) {
