+++ src/rendezvous/RendezvousIntent.ts
@@ -14,4 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-export * from './MSC3886SimpleHttpRendezvousTransport';
+export enum RendezvousIntent {
+  LOGIN_ON_NEW_DEVICE = "login.start",
+  RECIPROCATE_LOGIN_ON_EXISTING_DEVICE = "login.reciprocate",
+}
