+++ src/rendezvous/RendezvousTransport.ts
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { RendezvousFailureListener, RendezvousFailureReason } from "./cancellationReason";
+import { RendezvousFailureListener, RendezvousFailureReason } from ".";
 
 export interface RendezvousTransportDetails {
   type: string;
